personal-teacher-langchain/requirements.txt
python-dotenv
langgraph
typing
langchain_core
langchain_openai
langchain_community
colorama
